Platform/FLServerFastAPI/pyproject.toml
[project]
name = "fl-server-fast-api"
version = "0.1.0"
description = ""
authors = [
    {name = "José Antonio Peregrina Pérez",email = "jsperegrina1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "authlib (>=1.6.4,<2.0.0)",
    "websockets (>=15.0.1)",
    "python-keycloak (>=3.3.0)",
    "pymongo (>=4.8.0)",
    "pydantic (>= 2.10.6)",
    "fastapi-keycloak-middleware (>=1.3.0,<2.0.0)",
    "flwr (==1.9.0)",
    "tensorflow (== 2.16.1)",
    "pandas (>=1.3.0, <=2.2)",
    "keras (==3.3.3)",
    "numpy (>=1.24.4, <=1.27)",
    "scikit-learn (==1.4.2)",
    "great-expectations (==1.2.6)",
    "optuna (==3.2.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "xgboost (>=2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose (>= 3.5.0)"
]

[tool.poetry]
packages = [{include = "fl-server-fast-api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
